+++ src/legato/api/mkBuild.ts
@@ -1,10 +1,10 @@
 'use strict';
 
-import * as vscode from 'vscode';
+import { pathExists } from 'fs-extra';
 import * as path from 'path';
-import { LegatoFileExtension } from './core';
-import { ModelElement } from '../../commons/model';
-import { DisposableBag } from '../../commons/manager';
+import { Command } from '../../commons/identifiers';
+import { CommandRegister } from '../../commons/manager';
+import { LegatoFileExtension, LegatoManager } from './core';
 
 /**
  * Build tools
@@ -17,27 +17,23 @@ export const enum LegatoMkTools {
 /**
  * Manage mktools commands for build
  */
-export class MkBuildManager extends DisposableBag {
-    // Exposed model
-    public readonly mkTool = new ModelElement<LegatoMkTools | undefined>("legato.mkbuild.mktool", this);
-    public readonly buildCommand = new ModelElement<string | undefined>("legato.mkbuild.cmd.build", this);
-    public readonly buildAndInstallCommand = new ModelElement<string | undefined>("legato.mkbuild.cmd.buildandinstall", this);
+export class MkBuildManager extends CommandRegister {
 
     /**
-     * We need current def file to get the right mktool
+     * On build demand, we need legatoManager to get the right mktool based on def file
      */
-    public constructor(defFile: ModelElement<vscode.Uri | undefined>) {
+    public constructor(private readonly legatoManager: LegatoManager) {
         super();
-        defFile.addDependency(this.mkTool, this.getMkTool, this);
-        this.mkTool.addDependency(this.buildCommand, this.getBuildCommand, this);
-        this.buildCommand.addDependency(this.buildAndInstallCommand, this.getBuildAndInstallCommand, this);
+        // enables the possibility to create custom launch with 'eval $(leaf env -q) && ${command:legato.build}'
+        this.createCommand(Command.LegatoBuildCommand, this.getBuildCommand, this);
     }
 
     /**
      * @param defFile the active def file
      * @returns the corresponding mktool
      */
-    private getMkTool(defFile: vscode.Uri | undefined): LegatoMkTools | undefined {
+    private async getMkTool(): Promise<LegatoMkTools | undefined> {
+        const defFile = await this.legatoManager.defFile.get();
         if (defFile) {
             switch (path.extname(defFile.fsPath)) {
                 case LegatoFileExtension.sdef:
@@ -53,10 +49,24 @@ export class MkBuildManager extends DisposableBag {
      * @param mktool the current mktool
      * @returns the corresponding build command
      */
-    public getBuildCommand(mktool: LegatoMkTools | undefined): string | undefined {
+    public async getBuildCommand(): Promise<string | undefined> {
+        const mktool = await this.getMkTool();
         if (mktool) {
-            return `${mktool} -t \${LEGATO_TARGET} \${LEGATO_DEF_FILE}`;
+            let command = `${mktool} ${this.legatoManager.defFile.getEvalExpression()} -s components -t ${this.legatoManager.legatoTarget.getEvalExpression()}`;
+
+            // object dir arg complete if not empty
+            command = await this.legatoManager.legatoObjectDir.get() ? command.concat(" -w ", this.legatoManager.legatoObjectDir.getEvalExpression()) : command;
+
+            // output dir arg complete if not empty
+            command = await this.legatoManager.legatoOutputDir.get() ? command.concat(" -o ", this.legatoManager.legatoOutputDir.getEvalExpression()) : command;
+
+            // debug directory
+            command = await this.legatoManager.legatoDebugDir.get() ? command.concat(" -d ", this.legatoManager.legatoDebugDir.getEvalExpression()) : command;
+
+            return command;
         }
+
+        // unavailable build command when no def file is present
         return undefined;
     }
 
@@ -64,10 +74,42 @@ export class MkBuildManager extends DisposableBag {
      * @param buildCommand the current build command
      * @returns the corresponding build and install command
      */
-    public getBuildAndInstallCommand(buildCommand: string | undefined): string | undefined {
+    public async getBuildAndInstallCommand(): Promise<string | undefined> {
+        const buildCommand = await this.getBuildCommand();
         if (buildCommand) {
-            return `${buildCommand} && update $(basename \${LEGATO_DEF_FILE%.*def}).$LEGATO_TARGET.update`;
+            return `${buildCommand} && update ${this.legatoManager.legatoUpdateFile.getEvalExpression()}`;
         }
         return undefined;
     }
+
+    /**
+     * @returns clean command if the environment variables specifying the artifacts to remove are set
+     */
+    public async getCleanCommand() {
+        const artifacts: Array<string> = [];
+        if (await this.legatoManager.legatoDebugDir.get()) {
+            artifacts.push(this.legatoManager.legatoDebugDir.getEvalExpression());
+        }
+        if (await this.legatoManager.legatoObjectDir.get()) {
+            artifacts.push(this.legatoManager.legatoObjectDir.getEvalExpression());
+        }
+        if (await this.legatoManager.legatoUpdateFile.get()) {
+            artifacts.push(this.legatoManager.legatoUpdateFile.getEvalExpression());
+        }
+
+        return artifacts.length > 0 ? "rm -rf ".concat(...artifacts.map(value => value.concat(" "))) : undefined;
+    }
+
+    /**
+     * If the update file exists, generate image
+     */
+    public async generateImage() {
+        //if the .update file exists, the generate image command is provided
+        const updateFile = await this.legatoManager.legatoUpdateFile.getResolvedPath();
+        if (updateFile && await pathExists(updateFile)) {
+            return `systoimg ${this.legatoManager.legatoTarget.getEvalExpression()} ${this.legatoManager.legatoUpdateFile.getEvalExpression()} ${this.legatoManager.legatoObjectDir.getEvalExpression()}/image`;
+        }
+        return undefined;
+    }
+
 }